+++ src/app/user/dashboard/page.tsx
@@ -1,6 +1,6 @@
 import { api } from "~/trpc/server";
-import ShippingHistoryTable from "./components/ShippingHistoryTable";
-import Card from "./components/Card";
+import ShippingHistoryTable from "../../components/ShippingHistoryTable";
+import Card from "../../components/Card";
 import { DocumentPlusIcon, BanknotesIcon, ClipboardDocumentListIcon } from "@heroicons/react/24/solid";
 
 export default async function Dashboard() {
