+++ src/app/user/layout.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from "next";
-import Sidebar from "./components/Sidebar";
-import Header from "./components/Header";
+import Sidebar from "../components/Sidebar";
+import Header from "../components/Header";
 
 export const metadata: Metadata = {
   title: "User Site",
